Conflicts:
	SobekCM_Library/Database/SobekCM_Database.cs
	SobekCM_Library/HTML/Statistics_HtmlSubwriter.cs
	SobekCM_Library/ItemViewer/Viewers/QC_ItemViewer.cs
	SobekCM_Library/SobekCM_Library.csproj

+++ SobekCM_Engine_Library/Items/SobekCM_METS_Based_ItemBuilder.cs
@@ -638,11 +638,14 @@ namespace SobekCM.Engine_Library.Items
 
 			pageseq = 0;
 			List<Page_TreeNode> pages_encountered = new List<Page_TreeNode>();
-			foreach (abstract_TreeNode rootNode in Package_To_Finalize.Divisions.Physical_Tree.Roots)
-			{
-				recurse_through_nodes(Package_To_Finalize, rootNode, pages_encountered);
-			}
-			Package_To_Finalize.Web.Static_PageCount = pages_encountered.Count;
+		    if (!Package_To_Finalize.Behaviors.Dark_Flag)
+		    {
+		        foreach (abstract_TreeNode rootNode in Package_To_Finalize.Divisions.Physical_Tree.Roots)
+		        {
+		            recurse_through_nodes(Package_To_Finalize, rootNode, pages_encountered);
+		        }
+		    }
+		    Package_To_Finalize.Web.Static_PageCount = pages_encountered.Count;
 			Package_To_Finalize.Web.Static_Division_Count = divseq;
 
 			// Make sure no icons were retained from the METS file itself
@@ -942,10 +945,11 @@ namespace SobekCM.Engine_Library.Items
 			}
 
 			// Step through each download and make sure it is fully built
-			if (Package_To_Finalize.Divisions.Download_Tree.Has_Files)
+			if (( !Package_To_Finalize.Behaviors.Dark_Flag ) && ( Package_To_Finalize.Divisions.Download_Tree.Has_Files))
 			{
 				string ead_file = String.Empty;
 				int pdf_download = 0;
+			    int video_download = 0;
 				string pdf_download_url = String.Empty;
 				List<abstract_TreeNode> downloadPages = Package_To_Finalize.Divisions.Download_Tree.Pages_PreOrder;
 			    string xsl = String.Empty;
@@ -1006,8 +1010,57 @@ namespace SobekCM.Engine_Library.Items
                                     }
                                 }
                                 break;
+
+                            
+                            case "WEBM":
+                            case "OGG":
+                            case "MP4":
+                            //case "AVI":
+                            //case "WMV":
+                            //case "MPG":
+                            //case "MOV":
+                            //case "FLV":
+                            //case "VOB":
+                            //case "WAV":
+                            //case "OGM":
+                            //case "MKV":
+                                video_download++;
+                                download_handled = true;
+                                break;
                         }
                     }
+
+                    // Check for video files
+				    if ((!download_handled) && ( downloadPage.Files != null ))
+				    {
+				        foreach (SobekCM_File_Info thisFileInfo in downloadPage.Files)
+				        {
+                            string extension = thisFileInfo.File_Extension;
+
+                            // Was this an EAD page?
+                            switch (extension)
+                            {
+                                case "WEBM":
+                                case "OGG":
+                                case "MP4":
+                                    //case "AVI":
+                                    //case "WMV":
+                                    //case "MPG":
+                                    //case "MOV":
+                                    //case "FLV":
+                                    //case "VOB":
+                                    //case "WAV":
+                                    //case "OGM":
+                                    //case "MKV":
+                                    video_download++;
+                                    download_handled = true;
+                                    break;
+                            }
+
+                            if (download_handled)
+                                break;
+				        }
+				    }
 
 					// Step through each download file
 					if (!download_handled)
@@ -1030,6 +1083,7 @@ namespace SobekCM.Engine_Library.Items
                                     pdf_download_url = thisFile.System_Name;
                                 }
 							}
+
 						}
 					}
 				}
@@ -1069,6 +1123,11 @@ namespace SobekCM.Engine_Library.Items
 				{
                     Package_To_Finalize.Behaviors.Add_View(View_Enum.PDF).FileName = pdf_download_url;
 				}
+
+			    if (video_download > 0)
+			    {
+			        Package_To_Finalize.Behaviors.Add_View(View_Enum.VIDEO);
+			    }
 			}
 			else
 			{
@@ -1086,14 +1145,14 @@ namespace SobekCM.Engine_Library.Items
 			}
 
 			// Look for the HTML type views next, and possible set some defaults
-			if (viewsFromDb.ContainsKey(View_Enum.HTML))
+            if ((!Package_To_Finalize.Behaviors.Dark_Flag) && viewsFromDb.ContainsKey(View_Enum.HTML))
 			{
 				Package_To_Finalize.Behaviors.Add_View(viewsFromDb[View_Enum.HTML]);
 				viewsFromDb.Remove(View_Enum.HTML);
 			}
 
 			// Copy the TEI flag
-			if (viewsFromDb.ContainsKey(View_Enum.TEI))
+            if ((!Package_To_Finalize.Behaviors.Dark_Flag) && viewsFromDb.ContainsKey(View_Enum.TEI))
 			{
 				Package_To_Finalize.Behaviors.Add_View(viewsFromDb[View_Enum.TEI]);
 				viewsFromDb.Remove(View_Enum.TEI);
@@ -1117,7 +1176,7 @@ namespace SobekCM.Engine_Library.Items
 			}
 
 			// Finally, add all the ITEM VIEWS
-			if ((Package_To_Finalize.Web.Pages_By_Sequence != null) && (Package_To_Finalize.Web.Pages_By_Sequence.Count > 0))
+            if ((!Package_To_Finalize.Behaviors.Dark_Flag) && (Package_To_Finalize.Web.Pages_By_Sequence != null) && (Package_To_Finalize.Web.Pages_By_Sequence.Count > 0))
 			{
                 // Look for the RELATED IMAGES view next
                 if (viewsFromDb.ContainsKey(View_Enum.RELATED_IMAGES))
